Refactor view packages
