+++ app/connectors/VatApiConnector.scala
@@ -21,8 +21,8 @@ import javax.inject.{Inject, Singleton}
 
 import config.AppConfig
 import connectors.httpParsers.VatReturnObligationsHttpParser.HttpGetResult
-import models.VatReturnObligation.Status
-import models.VatReturnObligations
+import models.obligations.VatReturnObligation.Status
+import models.obligations.VatReturnObligations
 import play.api.Logger
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.bootstrap.http.HttpClient
