+++ app/models/obligations/Obligation.scala
@@ -14,19 +14,12 @@
  * limitations under the License.
  */
 
-package models
+package models.obligations
 
 import java.time.LocalDate
 
-import play.api.libs.json.{Format, Json}
+trait Obligation {
 
-case class Payment(end: LocalDate,
-                   due: LocalDate,
-                   outstandingAmount: BigDecimal,
-                   status: String,
-                   periodKey: String) extends Obligation
+  def due: LocalDate
 
-object Payment {
-
-  implicit val format: Format[Payment] = Json.format[Payment]
 }
