+++ app/views/templates/nextPaymentSection.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import models.Payment
+@import models.payments.Payment
 @import views.html.templates.formatters.dates._
 @import views.html.templates.formatters.money._
 @(payment: Option[Payment])(implicit messages: Messages)
