+++ app/views/vatDetails/details.scala.html
@@ -14,8 +14,10 @@
  * limitations under the License.
  *@
 
-@(user: models.User,
-  details: models.viewModels.VatDetailsModel,
+@import models._
+
+@(user: User,
+  details: VatDetailsModel,
   serviceInfo: Html = Html(""),
   tradingName: String)(implicit request: Request[_], messages: Messages, appConfig: config.AppConfig)
