+++ test/connectors/httpParsers/VatReturnObligationsHttpParserSpec.scala
@@ -19,8 +19,9 @@ package connectors.httpParsers
 import java.time.LocalDate
 
 import connectors.httpParsers.VatReturnObligationsHttpParser.VatReturnsReads
-import models.{VatReturnObligations, _}
+import models._
 import models.errors.{BadRequestError, MultipleErrors, ServerSideError, UnexpectedStatusError, UnknownError}
+import models.obligations.{VatReturnObligation, VatReturnObligations}
 import play.api.http.Status
 import play.api.libs.json.Json
 import uk.gov.hmrc.http.HttpResponse
