+++ test/models/PaymentsSpec.scala
@@ -18,6 +18,7 @@ package models
 
 import java.time.LocalDate
 
+import models.payments.{Payment, Payments}
 import play.api.libs.json.Json
 import uk.gov.hmrc.play.test.UnitSpec
