+++ test/views/templates/NextReturnSectionTemplateSpec.scala
@@ -18,7 +18,7 @@ package views.templates
 
 import java.time.LocalDate
 
-import models.VatReturnObligation
+import models.obligations.VatReturnObligation
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
 import views.ViewBaseSpec
